Better Property Encoding & Client Option Validation

+++ Source/HiveMQtt/MQTT5/Types/MQTT5Properties.cs
@@ -12,7 +12,7 @@ public class MQTT5Properties
 
     public byte? PayloadFormatIndicator { get; set; }
 
-    public int? MessageExpiryInterval { get; set; }
+    public UInt32? MessageExpiryInterval { get; set; }
 
     public string? ContentType { get; set; }
 
@@ -20,12 +20,12 @@ public class MQTT5Properties
 
     public byte[]? CorrelationData { get; set; }
 
-    public int? SubscriptionIdentifier { get; set; }
+    public uint? SubscriptionIdentifier { get; set; }
 
     /// <summary>
     /// Gets or sets a value that represents the session expiration duration in use by the MQTT broker.
     /// </summary>
-    public int? SessionExpiryInterval { get; set; }
+    public UInt32? SessionExpiryInterval { get; set; }
 
     /// <summary>
     /// Gets or sets the Client Identifier which was assigned by the Server because a zero length Client
@@ -39,7 +39,7 @@ public class MQTT5Properties
     /// If this value is set, this overrides any Keep Alive sent in the Connect request.
     /// </para>
     /// </summary>
-    public int? ServerKeepAlive { get; set; }
+    public UInt16? ServerKeepAlive { get; set; }
 
     public string? AuthenticationMethod { get; set; }
 
@@ -47,7 +47,7 @@ public class MQTT5Properties
 
     public byte? RequestProblemInformation { get; set; }
 
-    public int? WillDelayInterval { get; set; }
+    public UInt32? WillDelayInterval { get; set; }
 
     public byte? RequestResponseInformation { get; set; }
 
@@ -64,7 +64,7 @@ public class MQTT5Properties
     /// Gets or sets a value that indicates the maximum number of QoS 1 and QoS 2 publications that the
     /// MQTT broker is willing to process concurrently.
     /// </summary>
-    public int? ReceiveMaximum { get; set; }
+    public UInt16? ReceiveMaximum { get; set; }
 
     /// <summary>
     /// Gets or sets a value that indicates the highest value that the Server will accept as a Topic Alias sent by the Client.
@@ -73,15 +73,15 @@ public class MQTT5Properties
     /// A value of 0 indicates that the Server does not accept any Topic Aliases on this connection.
     /// </para>
     /// </summary>
-    public int? TopicAliasMaximum { get; set; }
+    public UInt16? TopicAliasMaximum { get; set; }
 
-    public int? TopicAlias { get; set; }
+    public UInt16? TopicAlias { get; set; }
 
     /// <summary>
     /// Gets or sets a value that indicates the maximum "Quality of Service" level that the MQTT
     /// broker is willing to accept.
     /// </summary>
-    public int? MaximumQoS { get; set; }
+    public UInt16? MaximumQoS { get; set; }
 
     /// <summary>
     /// Gets or sets a value that indicates whether the MQTT broker supports Retained Messages.
@@ -97,7 +97,7 @@ public class MQTT5Properties
     /// Gets or sets a value that indicates the maximum packet size that the MQTT broker is willing
     /// accept.
     /// </summary>
-    public Int32? MaximumPacketSize { get; set; }
+    public UInt32? MaximumPacketSize { get; set; }
 
     /// <summary>
     /// Gets or sets a value indicating whether the Server supports Wildcard Subscriptions.
